+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- `WalletService` will pass through messages to event listeners as `message` in the callback data
+
 ## 0.8.0
 
 ### Added